Use DespawnVfx for AmmoHit
